requirements.txt
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0" \
    --hash=sha256:a8df96034aae6d2d50a4ebe8216326c61c3eb64836776504fcca410e5937a3ba \
    --hash=sha256:f5971a9226b701070a4bf2c38c89e5a3f0d64de8debda981d1db98583009122a
